update feature test, add forms to cart view, update cart model and controller
